Struct tsukuyomi_juniper::GraphQLRequest

pub struct GraphQLRequest(_);

A wrapper around an incoming GraphQL request from a client.

Methods

impl GraphQLRequest

pub fn single(request: GraphQLRequest) -> GraphQLRequest

Creates a single GraphQL request.

pub fn batches<I>(iter: I) -> GraphQLRequest where
I: IntoIterator<Item = GraphQLRequest>,

Creates a batch GraphQL requests.

pub fn from_query(s: &str) -> Result<GraphQLRequest, Error>

Parses a query string into a single GraphQL request.

pub fn from_payload(payload: &[u8]) -> Result<GraphQLRequest, Error>

Parses a query string into a set of GraphQL request.

The provided payload must be a valid JSON data.

pub fn is_batch(&self) -> bool

Returns true if this request is a batch request.