Struct tsukuyomi_juniper::GraphQLState

pub struct GraphQLState<T, QueryT, MutationT> where
    QueryT: GraphQLType<Context = T>,
    MutationT: GraphQLType<Context = T>,  { /* fields omitted */ }

The main type containing all contextual values for processing GraphQL requests.

Methods

impl<T, QueryT, MutationT> GraphQLState<T, QueryT, MutationT> where
QueryT: GraphQLType<Context = T>,
MutationT: GraphQLType<Context = T>,

pub fn new(
context: T,
root_node: RootNode<'static, QueryT, MutationT>
) -> GraphQLState<T, QueryT, MutationT>

Creates a new GraphQLState from components.

pub fn context(&self) -> &T

Returns the reference to context in this value.

pub fn root_node(&self) -> &RootNode<'static, QueryT, MutationT>

Returns the reference to root node in this value.

pub fn execute(&self, request: GraphQLRequest) -> Execute<T, QueryT, MutationT>

Create a future for processing the execution of a GraphQL request.