Crate tsukuyomi_juniper

Components for integrating GraphQL endpoints into Tsukuyomi.

Structs

CaptureErrors

GraphQLRequest
The type representing a GraphQL request from the client.

GraphQLResponse
The type representing the result from the executing a GraphQL request.

Traits

Schema
A marker trait representing a root node of GraphQL schema.

Functions

capture_errors
Creates a ModifyHandler that catches the all kind of errors that the handler throws and converts them into GraphQL errors.

graphiql_source
Creates a handler function which returns a GraphiQL source.

request
Create an Extractor that parses the incoming request as GraphQL query.